Server-side pagination works.
